Update several things for latest master.
Enum repr seems to have changed.  PartialOrd changed as well, and initializer lists for structs apparently don't like p3 : 0 as c_float any more, as the type of 0 can't be immediately determined.

+++ src/network/ftp.rs
@@ -26,7 +26,7 @@ e* Permission is granted to anyone to use this software for any purpose,
 
 use std::mem;
 use std::c_str::CString;
-use libc::{size_t, c_int};
+use libc::size_t;
 
 use traits::Wrappable;
 use network::IpAddress;
@@ -53,106 +53,106 @@ pub enum Status {
     // expect another reply before proceeding with a new command
 
     /// Restart marker reply
-    RestartMarkerReply          = ffi::RESTARTMARKERREPLY,
+    RestartMarkerReply          = ffi::RESTARTMARKERREPLY as int,
     /// Service ready in N minutes
-    ServiceReadySoon            = ffi::SERVICEREADYSOON,
+    ServiceReadySoon            = ffi::SERVICEREADYSOON as int,
     /// Data connection already opened, transfer starting
-    DataConnectionAlreadyOpened = ffi::DATACONNECTIONALREADYOPENED,
+    DataConnectionAlreadyOpened = ffi::DATACONNECTIONALREADYOPENED as int,
     /// File status ok, about to open data connection
-    OpeningDataConnection       = ffi::OPENINGDATACONNECTION,
+    OpeningDataConnection       = ffi::OPENINGDATACONNECTION as int,
 
     // 2xx: the requested action has been successfully completed
 
     /// Command ok
-    Ok                          = ffi::OK as c_int,
+    Ok                          = ffi::OK as int,
     /// Command not implemented
-    PointlessCommand            = ffi::POINTLESSCOMMAND as c_int,
+    PointlessCommand            = ffi::POINTLESSCOMMAND as int,
     /// System status, or system help reply
-    SystemStatus                = ffi::SYSTEMSTATUS as c_int,
+    SystemStatus                = ffi::SYSTEMSTATUS as int,
     /// Directory status
-    DirectoryStatus             = ffi::DIRECTORYSTATUS as c_int,
+    DirectoryStatus             = ffi::DIRECTORYSTATUS as int,
     /// File status
-    FileStatus                  = ffi::FILESTATUS as c_int,
+    FileStatus                  = ffi::FILESTATUS as int,
     /// Help message
-    HelpMessage                 = ffi::HELPMESSAGE as c_int,
+    HelpMessage                 = ffi::HELPMESSAGE as int,
     /// NAME system type, where NAME is an official system name from the list in the Assigned Numbers document
-    SystemType                  = ffi::SYSTEMTYPE as c_int,
+    SystemType                  = ffi::SYSTEMTYPE as int,
     /// Service ready for new user
-    ServiceReady                = ffi::SERVICEREADY as c_int,
+    ServiceReady                = ffi::SERVICEREADY as int,
     /// Service closing control connection
-    ClosingConnection           = ffi::CLOSINGCONNECTION as c_int,
+    ClosingConnection           = ffi::CLOSINGCONNECTION as int,
     /// Data connection open, no transfer in progress
-    DataConnectionOpened        = ffi::DATACONNECTIONOPENED as c_int,
+    DataConnectionOpened        = ffi::DATACONNECTIONOPENED as int,
     /// Closing data connection, requested file action successful
-    ClosingDataConnection       = ffi::CLOSINGDATACONNECTION as c_int,
+    ClosingDataConnection       = ffi::CLOSINGDATACONNECTION as int,
     /// Entering passive mode
-    EnteringPassiveMode         = ffi::ENTERINGPASSIVEMODE as c_int,
+    EnteringPassiveMode         = ffi::ENTERINGPASSIVEMODE as int,
     /// User logged in, proceed. Logged out if appropriate
-    LoggedIn                    = ffi::LOGGEDIN as c_int,
+    LoggedIn                    = ffi::LOGGEDIN as int,
     /// Requested file action ok
-    FileActionOk                = ffi::FILEACTIONOK as c_int,
+    FileActionOk                = ffi::FILEACTIONOK as int,
     /// PATHNAME created
-    DirectoryOk                 = ffi::DIRECTORYOK as c_int,
+    DirectoryOk                 = ffi::DIRECTORYOK as int,
 
     // 3xx: the command has been accepted, but the requested action
     // is dormant, pending receipt of further information
     /// User name ok, need password
-    NeedPassword                = ffi::NEEDPASSWORD as c_int,
+    NeedPassword                = ffi::NEEDPASSWORD as int,
     /// Need account for login
-    NeedAccountToLogIn          = ffi::NEEDACCOUNTTOLOGIN as c_int,
+    NeedAccountToLogIn          = ffi::NEEDACCOUNTTOLOGIN as int,
     /// Requested file action pending further information
-    NeedInformation             = ffi::NEEDINFORMATION as c_int,
+    NeedInformation             = ffi::NEEDINFORMATION as int,
 
     // 4xx: the command was not accepted and the requested action did not take place,
     // but the error condition is temporary and the action may be requested again
 
     /// Service not available, closing control connection
-    ServiceUnavailable          = ffi::SERVICEUNAVAILABLE as c_int,
+    ServiceUnavailable          = ffi::SERVICEUNAVAILABLE as int,
     /// Can't open data connection
-    DataConnectionUnavailable   = ffi::DATACONNECTIONUNAVAILABLE as c_int,
+    DataConnectionUnavailable   = ffi::DATACONNECTIONUNAVAILABLE as int,
     /// Connection closed, transfer aborted
-    TransferAborted             = ffi::TRANSFERABORTED as c_int,
+    TransferAborted             = ffi::TRANSFERABORTED as int,
     /// Requested file action not taken
-    FileActionAborted           = ffi::FILEACTIONABORTED as c_int,
+    FileActionAborted           = ffi::FILEACTIONABORTED as int,
     /// Requested action aborted, local error in processing
-    LocalError                  = ffi::LOCALERROR as c_int,
+    LocalError                  = ffi::LOCALERROR as int,
     /// Requested action not taken; insufficient storage space in system, file unavailable
-    InsufficientStorageSpace    = ffi::INSUFFICIENTSTORAGESPACE as c_int,
+    InsufficientStorageSpace    = ffi::INSUFFICIENTSTORAGESPACE as int,
 
     // 5xx: the command was not accepted and
     // the requested action did not take place
     /// Syntax error, command unrecognized
-    CommandUnknown              = ffi::COMMANDUNKNOWN as c_int,
+    CommandUnknown              = ffi::COMMANDUNKNOWN as int,
     /// Syntax error in parameters or arguments
-    ParametersUnknown           = ffi::PARAMETERSUNKNOWN as c_int,
+    ParametersUnknown           = ffi::PARAMETERSUNKNOWN as int,
     /// Command not implemented
-    CommandNotImplemented       = ffi::COMMANDNOTIMPLEMENTED as c_int,
+    CommandNotImplemented       = ffi::COMMANDNOTIMPLEMENTED as int,
     /// Bad sequence of commands
-    BadCommandSequence          = ffi::BADCOMMANDSEQUENCE as c_int,
+    BadCommandSequence          = ffi::BADCOMMANDSEQUENCE as int,
     /// Command not implemented for that parameter
-    ParameterNotImplemented     = ffi::PARAMETERNOTIMPLEMENTED as c_int,
+    ParameterNotImplemented     = ffi::PARAMETERNOTIMPLEMENTED as int,
     /// Not logged in
-    NotLoggedIn                 = ffi::NOTLOGGEDIN as c_int,
+    NotLoggedIn                 = ffi::NOTLOGGEDIN as int,
     /// Need account for storing files
-    NeedAccountToStore          = ffi::NEEDACCOUNTTOSTORE as c_int,
+    NeedAccountToStore          = ffi::NEEDACCOUNTTOSTORE as int,
     /// Requested action not taken, file unavailable
-    FileUnavailable             = ffi::FILEUNAVAILABLE as c_int,
+    FileUnavailable             = ffi::FILEUNAVAILABLE as int,
     /// Requested action aborted, page type unknown
-    PageTypeUnknown             = ffi::PAGETYPEUNKNOWN as c_int,
+    PageTypeUnknown             = ffi::PAGETYPEUNKNOWN as int,
     /// Requested file action aborted, exceeded storage allocation
-    NotEnoughMemory             = ffi::NOTENOUGHMEMORY as c_int,
+    NotEnoughMemory             = ffi::NOTENOUGHMEMORY as int,
     /// Requested action not taken, file name not allowed
-    FilenameNotAllowed          = ffi::FILENAMENOTALLOWED as c_int,
+    FilenameNotAllowed          = ffi::FILENAMENOTALLOWED as int,
 
     // 10xx: SFML custom codes
     /// Response is not a valid FTP one
-    InvalidResponse             = ffi::INVALIDRESPONSE as c_int,
+    InvalidResponse             = ffi::INVALIDRESPONSE as int,
     /// Connection with server failed
-    ConnectionFailed            = ffi::CONNECTIONFAILED as c_int,
+    ConnectionFailed            = ffi::CONNECTIONFAILED as int,
     /// Connection with server closed
-    ConnectionClosed            = ffi::CONNECTIONCLOSED as c_int,
+    ConnectionClosed            = ffi::CONNECTIONCLOSED as int,
     /// Invalid file to upload / download
-    InvalidFile                 = ffi::INVALIDFILE as c_int
+    InvalidFile                 = ffi::INVALIDFILE as int
 }
 
 /// The FTP client
